refactor(page): extract file/drop handling + export-thumbnail options (phase 1, batch 4)
Move file/extension/drop-handling helpers (plus the private plugin
file-type derivations) into features/import-export/fileHandling.ts, and
the export-thumbnail render options into
features/export/exportThumbnailOptions.ts. Verbatim move; call sites
re-import by name; no behavior change.

page.tsx: 22,356 -> 22,157 lines (-976 total across phase 1 batches
1-4; 15 helper/type/const modules extracted). Verified via
`tsc --noEmit` (zero errors in any touched file).
