+++ src/app/page.tsx
@@ -149,6 +149,28 @@ import {
   type HomeSupportCollectionsSnapshot,
   type HomeKickstandCollectionsSnapshot,
 } from '@/features/supports/supportSnapshotHelpers';
+import {
+  EXPORT_THUMBNAIL_RENDER_OPTIONS_STORAGE_KEY,
+  DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS,
+  resolveInitialExportThumbnailRenderOptions,
+  type ExportThumbnailRenderOptions,
+} from '@/features/export/exportThumbnailOptions';
+import {
+  PLUGIN_IMPORT_WARNING_DISMISSED_STORAGE_KEY,
+  getFileExtension,
+  getFileNameFromPath,
+  isDragonfruitTempArtifactPath,
+  isSupportedPrepareDropName,
+  getDroppedFileMimeType,
+  isSceneFileName,
+  normalizeActiveVoxlScenePath,
+  extractTauriDroppedPaths,
+  isLikelyFileDragPayload,
+  getPrepareDropSupportStateFromDataTransfer,
+  buildDroppedFilesSignature,
+  type LaunchSceneFileEntry,
+  type SceneFileHandoffPayload,
+} from '@/features/import-export/fileHandling';
 import { getPluginSceneOverlayLoader } from '@/features/plugins/pluginRegistry';
 import {
   type HullCacheEntry,
@@ -493,34 +515,6 @@ if (typeof window !== 'undefined') {
   installReactDevtoolsSemverGuard();
 }
 
-type ExportThumbnailRenderOptions = {
-  includeGradient: boolean;
-  includeBuildPlate: boolean;
-  includeGrid: boolean;
-  centerOnModel: boolean;
-};
-
-const EXPORT_THUMBNAIL_RENDER_OPTIONS_STORAGE_KEY = 'dragonfruit.slicing.thumbnailRenderOptions';
-const DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS: ExportThumbnailRenderOptions = {
-  includeGradient: false,
-  includeBuildPlate: false,
-  includeGrid: false,
-  centerOnModel: true,
-};
-
-const PLUGIN_SCENE_FILE_TYPES = GENERATED_BUILTIN_COMPLEX_PLUGIN_DEFINITIONS.flatMap(
-  (def) => (def.fileTypes ?? []).filter((ft) => ft.isSceneFile),
-);
-const PLUGIN_ALL_FILE_TYPES = GENERATED_BUILTIN_COMPLEX_PLUGIN_DEFINITIONS.flatMap(
-  (def) => def.fileTypes ?? [],
-);
-const PREPARE_DROP_EXTENSIONS = new Set([
-  '.stl', '.obj', '.3mf', '.voxl',
-  ...PLUGIN_ALL_FILE_TYPES.map((ft) => ft.fileExtension),
-]);
-const PLUGIN_IMPORT_WARNING_DISMISSED_STORAGE_KEY =
-  PLUGIN_SCENE_FILE_TYPES.find((ft) => ft.fileExtension === '.lys')?.importWarning?.storageKey
-  ?? 'dragonfruit.lysImportWarningDismissed';
 const COLD_START_SCENE_HANDOFF_DELAY_MS = 1150;
 const REMOTE_OFFLINE_LAYER_HEIGHT_GLOBAL_STORAGE_KEY = 'dragonfruit.slicing.remoteOfflineLayerHeightMm';
 const REMOTE_OFFLINE_LAYER_HEIGHT_CHANGED_EVENT = 'dragonfruit:slicing-remote-offline-layer-height-changed';
@@ -573,195 +567,6 @@ function createModelTransformKey(modelId: string, transform: ModelTransform): st
     transform.scale.y.toFixed(6),
     transform.scale.z.toFixed(6),
   ].join('|');
-}
-
-function getFileExtension(name: string): string {
-  const trimmed = name.trim().toLowerCase();
-  const dotIndex = trimmed.lastIndexOf('.');
-  if (dotIndex < 0 || dotIndex === trimmed.length - 1) return '';
-  return trimmed.slice(dotIndex);
-}
-
-function getFileNameFromPath(path: string): string {
-  const normalized = path.replace(/\\/g, '/');
-  const parts = normalized.split('/').filter(Boolean);
-  return parts[parts.length - 1] ?? path;
-}
-
-function isDragonfruitTempArtifactPath(path: string | null | undefined): boolean {
-  if (typeof path !== 'string') return false;
-  const trimmed = path.trim();
-  if (!trimmed) return false;
-  const name = getFileNameFromPath(trimmed).toLowerCase();
-  return name.startsWith('dragonfruit-slice-');
-}
-
-function isSupportedPrepareDropName(name: string): boolean {
-  return PREPARE_DROP_EXTENSIONS.has(getFileExtension(name));
-}
-
-function getDroppedFileMimeType(name: string): string {
-  const ext = getFileExtension(name);
-  if (ext === '.stl') return 'model/stl';
-  if (ext === '.obj') return 'model/obj';
-  if (ext === '.3mf') return 'model/3mf';
-  if (ext === '.voxl') return 'application/json';
-  const pluginType = PLUGIN_ALL_FILE_TYPES.find((ft) => ft.fileExtension === ext);
-  return pluginType?.mimeType ?? 'application/octet-stream';
-}
-
-function isSceneFileName(name: string): boolean {
-  const ext = getFileExtension(name);
-  if (ext === '.voxl') return true;
-  return PLUGIN_SCENE_FILE_TYPES.some((ft) => ft.fileExtension === ext);
-}
-
-function normalizeActiveVoxlScenePath(path: string | null | undefined): string | null {
-  if (typeof path !== 'string') return null;
-  const trimmed = path.trim();
-  if (!trimmed) return null;
-  return getFileExtension(trimmed) === '.voxl' ? trimmed : null;
-}
-
-type LaunchSceneFileEntry = {
-  path: string;
-  name: string;
-};
-
-type SceneFileHandoffPayload = {
-  paths?: string[];
-  source?: string;
-};
-
-function extractTauriDroppedPaths(payload: unknown): string[] {
-  const isStringArray = (value: unknown): value is string[] => (
-    Array.isArray(value) && value.every((item) => typeof item === 'string')
-  );
-
-  if (isStringArray(payload)) {
-    return payload;
-  }
-
-  if (payload && typeof payload === 'object' && 'paths' in payload) {
-    const candidate = (payload as { paths?: unknown }).paths;
-    if (isStringArray(candidate)) {
-      return candidate;
-    }
-  }
-
-  return [];
-}
-
-function isLikelyFileDragPayload(dataTransfer: DataTransfer | null): boolean {
-  if (!dataTransfer) return false;
-  if ((dataTransfer.files?.length ?? 0) > 0) return true;
-  if (Array.from(dataTransfer.items ?? []).some((item) => item.kind === 'file')) return true;
-  if (Array.from(dataTransfer.types ?? []).includes('Files')) return true;
-  // Desktop runtime drags may not expose file metadata until drop.
-  return true;
-}
-
-function getPrepareDropSupportStateFromDataTransfer(dataTransfer: DataTransfer | null): 'supported' | 'unsupported' | 'unknown' {
-  if (!dataTransfer) return 'unknown';
-
-  const fileNames = new Set<string>();
-
-  const directFiles = Array.from(dataTransfer.files ?? []);
-  for (const file of directFiles) {
-    if (typeof file.name === 'string' && file.name.trim().length > 0) {
-      fileNames.add(file.name.trim());
-    }
-  }
-
-  const items = Array.from(dataTransfer.items ?? []);
-  for (const item of items) {
-    if (item.kind !== 'file') continue;
-    try {
-      const file = item.getAsFile();
-      if (file && typeof file.name === 'string' && file.name.trim().length > 0) {
-        fileNames.add(file.name.trim());
-      }
-
-      const webkitEntry = (item as DataTransferItem & {
-        webkitGetAsEntry?: () => { isFile?: boolean; name?: string } | null;
-      }).webkitGetAsEntry?.();
-      if (webkitEntry?.isFile && typeof webkitEntry.name === 'string' && webkitEntry.name.trim().length > 0) {
-        fileNames.add(webkitEntry.name.trim());
-      }
-    } catch {
-      // Some runtimes throw here during drag hover metadata probing.
-    }
-  }
-
-  const maybeExtractNameFromTextPath = (raw: string) => {
-    const trimmed = raw.trim();
-    if (!trimmed) return;
-
-    const firstLine = trimmed.split(/\r?\n/).find((line) => line.trim().length > 0)?.trim() ?? '';
-    if (!firstLine) return;
-
-    let normalized = firstLine;
-    if (normalized.startsWith('file://')) {
-      try {
-        normalized = decodeURIComponent(normalized.replace(/^file:\/\//, ''));
-      } catch {
-        normalized = normalized.replace(/^file:\/\//, '');
-      }
-    }
-
-    const name = getFileNameFromPath(normalized);
-    if (name.trim().length > 0) {
-      fileNames.add(name.trim());
-    }
-  };
-
-  try {
-    maybeExtractNameFromTextPath(dataTransfer.getData('text/uri-list'));
-    maybeExtractNameFromTextPath(dataTransfer.getData('text/plain'));
-  } catch {
-    // Ignore dataTransfer text extraction failures on restricted drag payloads.
-  }
-
-  if (fileNames.size === 0) {
-    return 'unknown';
-  }
-
-  const hasSupported = Array.from(fileNames).some((name) => isSupportedPrepareDropName(name));
-  return hasSupported ? 'supported' : 'unsupported';
-}
-
-function buildDroppedFilesSignature(files: File[]): string {
-  return files
-    .map((file) => `${file.name.trim().toLowerCase()}::${Number.isFinite(file.size) ? file.size : -1}`)
-    .sort((a, b) => a.localeCompare(b))
-    .join('|');
-}
-
-function resolveInitialExportThumbnailRenderOptions(): ExportThumbnailRenderOptions {
-  if (typeof window === 'undefined') return DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS;
-
-  try {
-    const raw = window.localStorage.getItem(EXPORT_THUMBNAIL_RENDER_OPTIONS_STORAGE_KEY);
-    if (!raw) return DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS;
-
-    const parsed = JSON.parse(raw) as Partial<ExportThumbnailRenderOptions>;
-    return {
-      includeGradient: typeof parsed.includeGradient === 'boolean'
-        ? parsed.includeGradient
-        : DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS.includeGradient,
-      includeBuildPlate: typeof parsed.includeBuildPlate === 'boolean'
-        ? parsed.includeBuildPlate
-        : DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS.includeBuildPlate,
-      includeGrid: typeof parsed.includeGrid === 'boolean'
-        ? parsed.includeGrid
-        : DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS.includeGrid,
-      centerOnModel: typeof parsed.centerOnModel === 'boolean'
-        ? parsed.centerOnModel
-        : DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS.centerOnModel,
-    };
-  } catch {
-    return DEFAULT_EXPORT_THUMBNAIL_RENDER_OPTIONS;
-  }
 }
 
 export default function Home() {
